+++ src/fl/audio/audio_detector.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "fl/ptr.h"
+
+namespace fl {
+
+class AudioContext;
+
+class AudioDetector {
+public:
+    virtual ~AudioDetector() = default;
+
+    virtual void update(shared_ptr<AudioContext> context) = 0;
+    virtual bool needsFFT() const { return false; }
+    virtual bool needsFFTHistory() const { return false; }
+    virtual const char* getName() const = 0;
+    virtual void reset() {}
+};
+
+} // namespace fl